Выносит функции для запросов api в отдельный файл
функции для запросов api вынесла в отдельный файл apiReqst в utils
Подправила расположение функций в apiServices
спиннер вынесла вместе с запросами к api
