+++ src/js/apiServices.js
@@ -1,51 +1,6 @@
-import axios from 'axios';
-import spinner from '../js/utils/spiner';
-// Для работы с API используем объект и его методы DataProccessing
-// keywordSearch(keyword) - для поиска по сключевому слову
-// getPopular() - получить список популярных фильмов
-// getNextPage(page) - возвращает данные для страницы page
+import { getPage, getGenres, getMovieById } from './utils/apiReqst';
 
-const API_KEY = '15ccc9a8c676c1c9b5477fb06b4d7b82';
-
-axios.defaults.baseURL = 'https://api.themoviedb.org/3/';
-
-// ------ Запросы к API ------
-
-// Сформировать путь запроса к популярным
-const getPopularPath = pageNum => {
-  spinner.stop();
-  return `movie/popular?api_key=${API_KEY}&language=en-US&page=${pageNum}&region=UA`;
-};
-
-// Сформировать путь запроса для поиска по ключевому слову
-const getKeywordPath = (keyword, pageNum) => {
-  return `search/movie?api_key=${API_KEY}&language=en-US&query=${keyword.replace(
-    ' ',
-    '+',
-  )}&page=${pageNum}&include_adult=false`;
-};
-
-// Получить страницу по номеру
-const getPage = (keyword, pageNum) => {
-  let url;
-  if (keyword === '') url = getPopularPath(pageNum);
-  else url = getKeywordPath(keyword, pageNum);
-  return axios.get(url).then(res => res.data);
-};
-
-// Получить перечень жанров фильмов
-const getGenres = () => {
-  const url = `/genre/movie/list?api_key=${API_KEY}&language=en-US`;
-  return axios.get(url).then(res => res.data);
-};
-
-// Получить данные о фильму по id
-export const getMovieById = id => {
-  const url = `movie/${id}?api_key=${API_KEY}`;
-  return axios.get(url).then(res => res.data);
-};
-
-//
+// Возвращает массив инфо фильмов по входному массиву ID
 export const getMovieByIdArray = idArray => {
   const promiseRes = new Promise((resolve, reject) => {
     resolve(
@@ -60,10 +15,6 @@ export const getMovieByIdArray = idArray => {
   return promiseRes;
 };
 
-// Константа кол-во фильмов на каждой странице от API
-const API_RESULTS_PER_PAGE = 20;
-
-let genresList = [];
 
 // ------ Функции для обработки массива жанров ------
 
@@ -109,6 +60,16 @@ function changeMovieRating(filteringArray) {
     item => (item.vote_average = item.vote_average.toFixed(1)),
   );
 }
+
+// Для работы с API используем объект и его методы DataProccessing
+// keywordSearch(keyword) - для поиска по сключевому слову
+// getPopular() - получить список популярных фильмов
+// getNextPage(page) - возвращает данные для страницы page
+
+// Константа кол-во фильмов на каждой странице от API
+const API_RESULTS_PER_PAGE = 20;
+
+let genresList = [];
 
 // Объект для формирования, отправки запроса и дальнейшей обработки данных
 class ApiRequest {
